auth-14
исправление замечаний ментора
(ДОМ отдельно от сервисов)
